Add Tests to Cabal project
Add some HSpec specification tests to the Cabal project that test the
Haq module that was moved from the root directory into `src` source
code.

Update the Haq module to be an actual module, and not a code sample,
with `module Haq where` at the beginning of the file. Then it can be
added as a compilable library amongst the other libraries / executables
within the `.cabal` file and require for the sample HSpec test.

== Notes:
- Add some binstubs to detail the commands used to build and run tests

== References:
- [How to write a Haskell program \- HaskellWiki]
  (https://wiki.haskell.org/How_to_write_a_Haskell_program#Add_some_tests)

- [testing \- How can I set up a simple test with Cabal? \- Stack
  Overflow]
  (https://stackoverflow.com/questions/1044555/how-can-i-set-up-a-simple-test-with-cabal)

- [3\.1\. Quickstart — Cabal <release> User's Guide]
  (https://www.haskell.org/cabal/users-guide/developing-packages.html#pkg-field-library-exposed-modules)