Add Entity typeclass example
Add Entity type-class example to demonstrate the purposes of type
classes in a Haskell application, even though a simple `data` type would
be fine in this Toy Robot application.

I would effectively describe a _"class"_ as either an abstract class or
interface.

The reason for this is that it contains all the expected function
prototypes that instances of the typeclass will contian, so that
multiple data types that share those common function prototypes do not
need to redeclare and instead inherit them, so that they can focus just
on their implementation.

This all seems very confusing without a good example.

== Notes:
- Also a `new ()` function is ambiguous and cannot be done within a
  class interface, since it needs to conform to one type, not multiple

== References:
- [OOP vs type classes \- HaskellWiki]
  (https://wiki.haskell.org/OOP_vs_type_classes)

- [7\.6\. Class and instances declarations]
  (https://downloads.haskell.org/~ghc/7.6.3/docs/html/users_guide/type-class-extensions.html)

- [class \- Haskell record syntax and type classes \- Stack Overflow]
  (https://stackoverflow.com/questions/1897306/haskell-record-syntax-and-type-classes)

- [Haskell/Classes and types \- Wikibooks, open books for an open world]
  (https://en.wikibooks.org/wiki/Haskell/Classes_and_types)

- [Yet Another Haskell Tutorial/Type advanced \- Wikibooks, open books
  for an open world]
  (https://en.wikibooks.org/wiki/Yet_Another_Haskell_Tutorial/Type_advanced#Instances)

- [haskell \- export from module \- Stack Overflow]
  (https://stackoverflow.com/questions/8983826/export-from-module)