Add Robot move function
Add a Robot move function to Robot module, it returns an updated Robot
data instance with the following changes, based off its previous state:

- increases `y` by 1, when facing NORTH
- decreases `y` by 1, when facing SOUTH
- increases `x` by 1, when facing EAST
- decreases `x` by 1, when facing WEST

Also, it does increase or decrease either `x` or `y` if the resulting
value is going to be less than 0 or greater than 4. As required by the
Toy Robot specification.