Refactor `run` to accept `args` parameter
Refactor the `run` command to accept a String List argument as a
parameter that will be passed to it from `main`, since that is where
`getArgs` should be and only called from, rather than it happening
elsewhere in the application via a global grab function.

This change also means that arguments can be feed into during tests,
however `stdin` still will not yet be accessible via this change.

== Notes:
- Remove unused `readCommandline` function that was achieved using just
  `getLine` in the User Loop function

- Also refactor `tulifyInput` to be without a `case of` statement

== References:
- [Haskell/Pattern matching \- Wikibooks, open books for an open world]
  (https://en.wikibooks.org/wiki/Haskell/Pattern_matching)